requirements.txt
pandas>=1.5
numpy>=1.24
scipy>=1.11
scikit-learn>=1.2
pyyaml>=6.0
openai>=1.0.0
tqdm
python-dotenv
